requirements.txt
anchor-exp>=0.0.2.0,<1
dalex>=1.5.0,<2
dice-ml==0.11
dill>=0.3.9,<4
Flask[async]
flask-cors>=3.0.13
gin-config==0.5.0
gunicorn==23.0.0
inflect==5.4.0
lime==0.2.0.1
llama-index-core==0.13.1
llama-index-instrumentation==0.4.0
llama-index-llms-openai==0.5.2
llama-index-utils-workflow==0.4.0
matplotlib==3.7.0
numpy>=1.23.5,<2.0.0
openai==1.99.6
orjson
pandas==1.5.3
pandas-stubs==1.2.0.53
pydantic==2.11.7
pytest==7.2.0
python-dateutil==2.8.2
pytz==2022.1
python-dotenv==1.0.1
requests==2.32.4
scikit-learn==1.0.2
scipy>=1.8.0,<2.0.0
shap
statsmodels==0.13.2
sklearn==0.0
tqdm==4.67.1
typing_extensions==4.12.2
transitions>=0.9.1,<1
seaborn==0.12.2
setuptools
